+++ src/syntax/iter.c
@@ -91,3 +91,26 @@ SyntaxToken GetNextToken(LineIterator *iter)
 
     Assert(false); // Unreachable
 }
+
+// Returns true if sequence was found, and also keeps iteration made to iter.
+// Otherwise iteration is reset to where it was.
+bool MatchSymbolSequence(LineIterator *iter, char *sequence)
+{
+    int prevPos = iter->pos; // Reset to this if comment not found
+    bool matched = true;
+
+    for (int i = 1; i < (int)strlen(sequence); i++)
+    {
+        SyntaxToken tok = GetNextToken(iter);
+        if (tok.eof || !tok.isSymbol || tok.word[0] != sequence[i])
+        {
+            matched = false;
+            break;
+        }
+    }
+
+    if (!matched)
+        iter->pos = prevPos;
+
+    return matched;
+}